change end edit command to break
